third_party/blink/web_tests/http/tests/shadowrealm/webexposed/global-interface-listing-shadow-realm-expected.txt

CONSOLE MESSAGE: This test logs exposed APIs from ShadowRealmGlobalScope
CONSOLE MESSAGE: [INTERFACES]
CONSOLE MESSAGE: interface AbortController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface AbortSignal : EventTarget
CONSOLE MESSAGE:     static method abort
CONSOLE MESSAGE:     static method any
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter aborted
CONSOLE MESSAGE:     getter onabort
CONSOLE MESSAGE:     getter reason
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method throwIfAborted
CONSOLE MESSAGE:     setter onabort
CONSOLE MESSAGE: interface AsyncDisposableStack
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter disposed
CONSOLE MESSAGE:     method Symbol(Symbol.asyncDispose)
CONSOLE MESSAGE:     method adopt
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method defer
CONSOLE MESSAGE:     method disposeAsync
CONSOLE MESSAGE:     method move
CONSOLE MESSAGE:     method use
CONSOLE MESSAGE: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CompressionStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CountQueuingStrategy
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter highWaterMark
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface CustomEvent : Event
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter detail
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method initCustomEvent
CONSOLE MESSAGE: interface DOMException
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     attribute ABORT_ERR
CONSOLE MESSAGE:     attribute DATA_CLONE_ERR
CONSOLE MESSAGE:     attribute DOMSTRING_SIZE_ERR
CONSOLE MESSAGE:     attribute HIERARCHY_REQUEST_ERR
CONSOLE MESSAGE:     attribute INDEX_SIZE_ERR
CONSOLE MESSAGE:     attribute INUSE_ATTRIBUTE_ERR
CONSOLE MESSAGE:     attribute INVALID_ACCESS_ERR
CONSOLE MESSAGE:     attribute INVALID_CHARACTER_ERR
CONSOLE MESSAGE:     attribute INVALID_MODIFICATION_ERR
CONSOLE MESSAGE:     attribute INVALID_NODE_TYPE_ERR
CONSOLE MESSAGE:     attribute INVALID_STATE_ERR
CONSOLE MESSAGE:     attribute NAMESPACE_ERR
CONSOLE MESSAGE:     attribute NETWORK_ERR
CONSOLE MESSAGE:     attribute NOT_FOUND_ERR
CONSOLE MESSAGE:     attribute NOT_SUPPORTED_ERR
CONSOLE MESSAGE:     attribute NO_DATA_ALLOWED_ERR
CONSOLE MESSAGE:     attribute NO_MODIFICATION_ALLOWED_ERR
CONSOLE MESSAGE:     attribute QUOTA_EXCEEDED_ERR
CONSOLE MESSAGE:     attribute SECURITY_ERR
CONSOLE MESSAGE:     attribute SYNTAX_ERR
CONSOLE MESSAGE:     attribute TIMEOUT_ERR
CONSOLE MESSAGE:     attribute TYPE_MISMATCH_ERR
CONSOLE MESSAGE:     attribute URL_MISMATCH_ERR
CONSOLE MESSAGE:     attribute VALIDATION_ERR
CONSOLE MESSAGE:     attribute WRONG_DOCUMENT_ERR
CONSOLE MESSAGE:     getter code
CONSOLE MESSAGE:     getter message
CONSOLE MESSAGE:     getter name
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface DecompressionStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface DisposableStack
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter disposed
CONSOLE MESSAGE:     method Symbol(Symbol.dispose)
CONSOLE MESSAGE:     method adopt
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method defer
CONSOLE MESSAGE:     method dispose
CONSOLE MESSAGE:     method move
CONSOLE MESSAGE:     method use
CONSOLE MESSAGE: interface ErrorEvent : Event
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter colno
CONSOLE MESSAGE:     getter error
CONSOLE MESSAGE:     getter filename
CONSOLE MESSAGE:     getter lineno
CONSOLE MESSAGE:     getter message
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface Event
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     attribute AT_TARGET
CONSOLE MESSAGE:     attribute BUBBLING_PHASE
CONSOLE MESSAGE:     attribute CAPTURING_PHASE
CONSOLE MESSAGE:     attribute NONE
CONSOLE MESSAGE:     getter bubbles
CONSOLE MESSAGE:     getter cancelBubble
CONSOLE MESSAGE:     getter cancelable
CONSOLE MESSAGE:     getter composed
CONSOLE MESSAGE:     getter currentTarget
CONSOLE MESSAGE:     getter defaultPrevented
CONSOLE MESSAGE:     getter eventPhase
CONSOLE MESSAGE:     getter returnValue
CONSOLE MESSAGE:     getter srcElement
CONSOLE MESSAGE:     getter target
CONSOLE MESSAGE:     getter timeStamp
CONSOLE MESSAGE:     getter type
CONSOLE MESSAGE:     method composedPath
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method initEvent
CONSOLE MESSAGE:     method preventDefault
CONSOLE MESSAGE:     method stopImmediatePropagation
CONSOLE MESSAGE:     method stopPropagation
CONSOLE MESSAGE:     setter cancelBubble
CONSOLE MESSAGE:     setter returnValue
CONSOLE MESSAGE: interface EventTarget
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method addEventListener
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method dispatchEvent
CONSOLE MESSAGE:     method removeEventListener
CONSOLE MESSAGE:     method when
CONSOLE MESSAGE: interface Float16Array : TypedArray
CONSOLE MESSAGE:     attribute BYTES_PER_ELEMENT
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface Observable
CONSOLE MESSAGE:     static method from
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method catch
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method drop
CONSOLE MESSAGE:     method every
CONSOLE MESSAGE:     method filter
CONSOLE MESSAGE:     method finally
CONSOLE MESSAGE:     method find
CONSOLE MESSAGE:     method first
CONSOLE MESSAGE:     method flatMap
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method inspect
CONSOLE MESSAGE:     method last
CONSOLE MESSAGE:     method map
CONSOLE MESSAGE:     method reduce
CONSOLE MESSAGE:     method some
CONSOLE MESSAGE:     method subscribe
CONSOLE MESSAGE:     method switchMap
CONSOLE MESSAGE:     method take
CONSOLE MESSAGE:     method takeUntil
CONSOLE MESSAGE:     method toArray
CONSOLE MESSAGE: interface PromiseRejectionEvent : Event
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter promise
CONSOLE MESSAGE:     getter reason
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface QuotaExceededError : DOMException
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter quota
CONSOLE MESSAGE:     getter requested
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface ReadableByteStreamController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter byobRequest
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method @@asyncIterator
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getReader
CONSOLE MESSAGE:     method pipeThrough
CONSOLE MESSAGE:     method pipeTo
CONSOLE MESSAGE:     method tee
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface ReadableStreamBYOBReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ReadableStreamBYOBRequest
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter view
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method respond
CONSOLE MESSAGE:     method respondWithNewView
CONSOLE MESSAGE: interface ReadableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface ReadableStreamDefaultReader
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     method cancel
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method read
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE: interface ShadowRealm
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method evaluate
CONSOLE MESSAGE:     method importValue
CONSOLE MESSAGE: interface ShadowRealmGlobalScope
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface Subscriber
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter active
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method addTeardown
CONSOLE MESSAGE:     method complete
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method next
CONSOLE MESSAGE: interface SuppressedError : Error
CONSOLE MESSAGE:     attribute message
CONSOLE MESSAGE:     attribute name
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TextDecoder
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter encoding
CONSOLE MESSAGE:     getter fatal
CONSOLE MESSAGE:     getter ignoreBOM
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method decode
CONSOLE MESSAGE: interface TextDecoderStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter encoding
CONSOLE MESSAGE:     getter fatal
CONSOLE MESSAGE:     getter ignoreBOM
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TextEncoder
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter encoding
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method encode
CONSOLE MESSAGE:     method encodeInto
CONSOLE MESSAGE: interface TextEncoderStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter encoding
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter readable
CONSOLE MESSAGE:     getter writable
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE: interface TransformStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method enqueue
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method terminate
CONSOLE MESSAGE: interface URL
CONSOLE MESSAGE:     static method canParse
CONSOLE MESSAGE:     static method parse
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter hash
CONSOLE MESSAGE:     getter host
CONSOLE MESSAGE:     getter hostname
CONSOLE MESSAGE:     getter href
CONSOLE MESSAGE:     getter origin
CONSOLE MESSAGE:     getter password
CONSOLE MESSAGE:     getter pathname
CONSOLE MESSAGE:     getter port
CONSOLE MESSAGE:     getter protocol
CONSOLE MESSAGE:     getter search
CONSOLE MESSAGE:     getter searchParams
CONSOLE MESSAGE:     getter username
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method toJSON
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE:     setter hash
CONSOLE MESSAGE:     setter host
CONSOLE MESSAGE:     setter hostname
CONSOLE MESSAGE:     setter href
CONSOLE MESSAGE:     setter password
CONSOLE MESSAGE:     setter pathname
CONSOLE MESSAGE:     setter port
CONSOLE MESSAGE:     setter protocol
CONSOLE MESSAGE:     setter search
CONSOLE MESSAGE:     setter username
CONSOLE MESSAGE: interface URLPattern
CONSOLE MESSAGE:     static method compareComponent
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter hasRegExpGroups
CONSOLE MESSAGE:     getter hash
CONSOLE MESSAGE:     getter hostname
CONSOLE MESSAGE:     getter password
CONSOLE MESSAGE:     getter pathname
CONSOLE MESSAGE:     getter port
CONSOLE MESSAGE:     getter protocol
CONSOLE MESSAGE:     getter search
CONSOLE MESSAGE:     getter username
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method exec
CONSOLE MESSAGE:     method generate
CONSOLE MESSAGE:     method test
CONSOLE MESSAGE: interface URLSearchParams
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter size
CONSOLE MESSAGE:     method @@iterator
CONSOLE MESSAGE:     method append
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method delete
CONSOLE MESSAGE:     method entries
CONSOLE MESSAGE:     method forEach
CONSOLE MESSAGE:     method get
CONSOLE MESSAGE:     method getAll
CONSOLE MESSAGE:     method has
CONSOLE MESSAGE:     method keys
CONSOLE MESSAGE:     method set
CONSOLE MESSAGE:     method sort
CONSOLE MESSAGE:     method toString
CONSOLE MESSAGE:     method values
CONSOLE MESSAGE: interface WritableStream
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter locked
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method getWriter
CONSOLE MESSAGE: interface WritableStreamDefaultController
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter signal
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE: interface WritableStreamDefaultWriter
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     getter closed
CONSOLE MESSAGE:     getter desiredSize
CONSOLE MESSAGE:     getter ready
CONSOLE MESSAGE:     method abort
CONSOLE MESSAGE:     method close
CONSOLE MESSAGE:     method constructor
CONSOLE MESSAGE:     method releaseLock
CONSOLE MESSAGE:     method write
CONSOLE MESSAGE: [NAMESPACES]
CONSOLE MESSAGE: namespace Temporal
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     attribute Now
CONSOLE MESSAGE:     method Duration
CONSOLE MESSAGE:     method Instant
CONSOLE MESSAGE:     method PlainDate
CONSOLE MESSAGE:     method PlainDateTime
CONSOLE MESSAGE:     method PlainMonthDay
CONSOLE MESSAGE:     method PlainTime
CONSOLE MESSAGE:     method PlainYearMonth
CONSOLE MESSAGE:     method ZonedDateTime
CONSOLE MESSAGE: namespace console
CONSOLE MESSAGE:     attribute @@toStringTag
CONSOLE MESSAGE:     method assert
CONSOLE MESSAGE:     method clear
CONSOLE MESSAGE:     method context
CONSOLE MESSAGE:     method count
CONSOLE MESSAGE:     method countReset
CONSOLE MESSAGE:     method debug
CONSOLE MESSAGE:     method dir
CONSOLE MESSAGE:     method dirxml
CONSOLE MESSAGE:     method error
CONSOLE MESSAGE:     method group
CONSOLE MESSAGE:     method groupCollapsed
CONSOLE MESSAGE:     method groupEnd
CONSOLE MESSAGE:     method info
CONSOLE MESSAGE:     method log
CONSOLE MESSAGE:     method profile
CONSOLE MESSAGE:     method profileEnd
CONSOLE MESSAGE:     method table
CONSOLE MESSAGE:     method time
CONSOLE MESSAGE:     method timeEnd
CONSOLE MESSAGE:     method timeLog
CONSOLE MESSAGE:     method timeStamp
CONSOLE MESSAGE:     method trace
CONSOLE MESSAGE:     method warn
CONSOLE MESSAGE: [GLOBAL OBJECT]
CONSOLE MESSAGE:     attribute globalThis
CONSOLE MESSAGE:     getter isSecureContext
CONSOLE MESSAGE:     method atob
CONSOLE MESSAGE:     method btoa
CONSOLE MESSAGE:     method gc
CONSOLE MESSAGE:     method queueMicrotask
CONSOLE MESSAGE:     method reportError
CONSOLE MESSAGE:     method structuredClone
